+++ src/main/java/org/folio/rest/impl/InstanceBatchSyncApi.java
@@ -11,6 +11,7 @@ import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.InstancesPost;
 import org.folio.rest.jaxrs.resource.InstanceStorageBatchSynchronous;
 import org.folio.services.instance.InstanceService;
+import org.folio.utils.InstanceUtils;
 
 public class InstanceBatchSyncApi implements InstanceStorageBatchSynchronous {
   @Validate
@@ -20,8 +21,10 @@ public class InstanceBatchSyncApi implements InstanceStorageBatchSynchronous {
                                                   Handler<AsyncResult<Response>> asyncResultHandler,
                                                   Context vertxContext) {
 
+    var instances = InstanceUtils.copyPropertiesToInstances(entity.getInstances());
+
     new InstanceService(vertxContext, okapiHeaders)
-      .createInstances(entity.getInstances(), upsert, true)
+      .createInstances(instances.getInstances(), upsert, true)
       .otherwise(cause -> respond500WithTextPlain(cause.getMessage()))
       .onComplete(asyncResultHandler);
   }