+++ src/main/java/org/folio/rest/impl/InstanceBatchSyncUnsafeApi.java
@@ -11,6 +11,7 @@ import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.InstancesPost;
 import org.folio.rest.jaxrs.resource.InstanceStorageBatchSynchronousUnsafe;
 import org.folio.services.instance.InstanceService;
+import org.folio.utils.InstanceUtils;
 
 public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchronousUnsafe {
   @Validate
@@ -19,8 +20,10 @@ public class InstanceBatchSyncUnsafeApi implements InstanceStorageBatchSynchrono
                                                         Handler<AsyncResult<Response>> asyncResultHandler,
                                                         Context vertxContext) {
 
+    var instances = InstanceUtils.copyPropertiesToInstances(entity.getInstances());
+
     new InstanceService(vertxContext, okapiHeaders)
-      .createInstances(entity.getInstances(), true, false)
+      .createInstances(instances.getInstances(), true, false)
       .otherwise(cause -> respond500WithTextPlain(cause.getMessage()))
       .onComplete(asyncResultHandler);
   }