+++ src/test/java/org/folio/rest/api/AsyncMigrationTest.java
@@ -19,6 +19,7 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.doReturn;
@@ -26,14 +27,18 @@ import static org.mockito.Mockito.spy;
 
 import io.vertx.core.Context;
 import io.vertx.core.json.JsonObject;
+import io.vertx.sqlclient.Row;
+import io.vertx.sqlclient.RowSet;
 import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.IntStream;
 import junitparams.JUnitParamsRunner;
+import lombok.SneakyThrows;
 import org.apache.commons.collections4.map.CaseInsensitiveMap;
 import org.folio.persist.AsyncMigrationJobRepository;
 import org.folio.rest.jaxrs.model.AsyncMigrationJob;
@@ -43,17 +48,35 @@ import org.folio.rest.jaxrs.model.AsyncMigrations;
 import org.folio.rest.jaxrs.model.EffectiveCallNumberComponents;
 import org.folio.rest.jaxrs.model.Processed;
 import org.folio.rest.jaxrs.model.Published;
+import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.persist.PostgresClientFuturized;
 import org.folio.rest.support.sql.TestRowStream;
 import org.folio.services.migration.async.AsyncMigrationContext;
 import org.folio.services.migration.async.ShelvingOrderMigrationJobRunner;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 @RunWith(JUnitParamsRunner.class)
 public class AsyncMigrationTest extends TestBaseWithInventoryUtil {
 
+  private static final String UPDATE_JSONB_WITH_PUB_PERIOD = """
+    UPDATE %s_mod_inventory_storage.instance
+    SET jsonb = jsonb || jsonb_set(jsonb, '{publicationPeriod}', jsonb_build_object('start', '1999', 'end', '2001'))
+    RETURNING id::text;
+    """;
+  private static final String SELECT_JSONB =
+    "SELECT jsonb FROM %s_mod_inventory_storage.instance";
+
   private final AsyncMigrationJobRepository repository = getRepository();
+
+  @SneakyThrows
+  @Before
+  public void beforeEach() {
+    StorageTestSuite.deleteAll(itemsStorageUrl(""));
+    StorageTestSuite.deleteAll(holdingsStorageUrl(""));
+    StorageTestSuite.deleteAll(instancesStorageUrl(""));
+  }
 
   private static Map<String, String> okapiHeaders() {
     return new CaseInsensitiveMap<>(Map.of(TENANT.toLowerCase(), TENANT_ID));
@@ -139,10 +162,65 @@ public class AsyncMigrationTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void canMigrateInstancePublicationPeriod() {
+    var numberOfRecords = 10;
+
+    IntStream.range(0, numberOfRecords).parallel().forEach(v ->
+      instancesClient.create(new JsonObject()
+        .put("title", "test" + v)
+        .put("source", "MARC")
+        .put("instanceTypeId", "30fffe0e-e985-4144-b2e2-1e8179bdb41f")));
+
+    var countDownLatch = new CountDownLatch(1);
+    var query = String.format(UPDATE_JSONB_WITH_PUB_PERIOD, TENANT_ID);
+    postgresClient(getContext(), okapiHeaders()).execute(query)
+      .onSuccess(event -> countDownLatch.countDown());
+    // check jsonb contains 'publicationPeriod' data
+    RowSet<Row> selectResult = runSql(String.format(SELECT_JSONB, TENANT_ID));
+
+    assertEquals(10, selectResult.rowCount());
+    JsonObject jsonbData = selectResult.iterator().next().toJson().getJsonObject("jsonb");
+    assertNull(jsonbData.getJsonObject("dates"));
+    assertNotNull(jsonbData.getJsonObject("publicationPeriod"));
+
+
+    await().atMost(5, SECONDS).until(() -> countDownLatch.getCount() == 0L);
+
+    var migrationJob = asyncMigration.postMigrationJob(new AsyncMigrationJobRequest()
+      .withMigrations(List.of("publicationPeriodMigration")));
+
+    await().atMost(25, SECONDS).until(() -> asyncMigration.getMigrationJob(migrationJob.getId())
+      .getJobStatus() == AsyncMigrationJob.JobStatus.COMPLETED);
+
+    var job = asyncMigration.getMigrationJob(migrationJob.getId());
+
+    assertThat(job.getPublished().stream().map(Published::getCount)
+      .mapToInt(Integer::intValue).sum(), is(numberOfRecords));
+    assertThat(job.getProcessed().stream().map(Processed::getCount)
+      .mapToInt(Integer::intValue).sum(), is(numberOfRecords));
+    assertThat(job.getJobStatus(), is(AsyncMigrationJob.JobStatus.COMPLETED));
+    assertThat(job.getSubmittedDate(), notNullValue());
+
+    // check that the 'publicationPeriod' data has been migrated to the 'dates' data
+    var selectQuery = String.format(SELECT_JSONB, TENANT_ID);
+    RowSet<Row> result = runSql(selectQuery);
+
+    assertEquals(10, result.rowCount());
+    JsonObject entry = result.iterator().next().toJson();
+    JsonObject jsonb = entry.getJsonObject("jsonb");
+    JsonObject dates = jsonb.getJsonObject("dates");
+    assertNotNull(dates);
+    assertNull(jsonb.getString("publicationPeriod"));
+    assertEquals("1999", dates.getString("date1"));
+    assertEquals("2001", dates.getString("date2"));
+    assertEquals("8fa6d067-41ff-4362-96a0-96b16ddce267", dates.getString("dateTypeId"));
+  }
+
+  @Test
   public void canGetAvailableMigrations() {
     AsyncMigrations migrations = asyncMigration.getMigrations();
     assertNotNull(migrations);
-    assertEquals(Integer.valueOf(2), migrations.getTotalRecords());
+    assertEquals(Integer.valueOf(3), migrations.getTotalRecords());
     assertEquals("itemShelvingOrderMigration", migrations.getAsyncMigrations().get(0).getMigrations().get(0));
   }
 
@@ -191,5 +269,14 @@ public class AsyncMigrationTest extends TestBaseWithInventoryUtil {
 
   private AsyncMigrationJobRepository getRepository() {
     return new AsyncMigrationJobRepository(getContext(), okapiHeaders());
+  }
+
+  @SneakyThrows
+  private RowSet<Row> runSql(String sql) {
+    return PostgresClient.getInstance(getVertx())
+      .execute(sql)
+      .toCompletionStage()
+      .toCompletableFuture()
+      .get(TIMEOUT, TimeUnit.SECONDS);
   }
 }